+++ src/pages/admin/MainPage/NoGroupSection/index.tsx
