+++ src/pages/admin/SelectWeekPage/WeeklyDetailSection/ClosedDetail.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+import { useQuery } from '@tanstack/react-query';
+import { getDailyWorkers } from 'apis/getSchedule';
+import DailyWorkersTemplate from 'components/DailyWorkers/DailyWorkersTemplate';
+import useWeekSelector from 'hooks/useWeekSelector';
+import { stringDateMove } from 'utils/stringDateMove';
+
+const ClosedDetail = ({ startWeekDate }: { startWeekDate: string }): JSX.Element => {
+  const { day, WeekBarComponent } = useWeekSelector(0);
+  const date = stringDateMove(startWeekDate, day);
+
+  const { data: scheduleRes } = useQuery(['getDailyWorkers', day], () => getDailyWorkers({ date: date }), {
+    suspense: true,
+  });
+
+  return (
+    <>
+      <WeekBarComponent />
+      <DailyWorkersTemplate dailyData={scheduleRes?.data.schedule} />
+    </>
+  );
+};
+
+export default ClosedDetail;
